refact(GenerateDAG): rename some variables, make methods private when it makes sense, add context files to image instead of node, tests to internal
