+++ pkg/dag/image.go
@@ -17,6 +17,7 @@ type Image struct {
 	ExtraTags         []string               `yaml:"extra_tags,flow,omitempty"`
 	Dockerfile        *dockerfile.Dockerfile `yaml:"dockerfile,omitempty"`
 	IgnorePatterns    []string               `yaml:"ignore_patterns,flow,omitempty"`
+	ContextFiles      []string               `yaml:"-"`
 	NeedsRebuild      bool                   `yaml:"-"`
 	NeedsTests        bool                   `yaml:"-"`
 	RetagDone         bool                   `yaml:"-"`
